app/requirement.txt
fastapi
uvicorn
pyjwt
zodb
python-multipart
jinja2